Parse integers as long long, use as per constraint

+++ ds/bitmap.h
@@ -10,13 +10,13 @@
 #define BITS_PER_WORD 64
 
 bitmapNode* newBitmapNode();
-bitmapNode* initBitmapNode(unsigned long, bool);
-bool getBit(bitmapNode*, unsigned long);
-bool setBit(bitmapNode*, unsigned long, bool);
-bool setbitUtility(void* , unsigned long);
-bool clearbitUtility(void* , unsigned long);
-bool getbitUtility(void* , unsigned long);
-inline unsigned long word_offset(unsigned long);
-inline size_t bit_offset(unsigned long);
+bitmapNode* initBitmapNode(long long, bool);
+bool getBit(bitmapNode*, long long);
+bool setBit(bitmapNode*, long long, bool);
+bool setbitUtility(void* , long long);
+bool clearbitUtility(void* , long long);
+bool getbitUtility(void* , long long);
+inline unsigned long word_offset(long long);
+inline size_t bit_offset(long long);
 
 #endif
